AssertResult -> Absinthe.Case.Assertions
