+++ test/graphql_specification/null_values/list_test.exs
@@ -1,6 +1,5 @@
 defmodule GraphQL.Specification.NullValues.ListTest do
   use Absinthe.Case, async: true
-  import AssertResult
 
   defmodule Schema do
     use Absinthe.Schema
