+++ test/lib/absinthe/execution/arguments_test.exs
@@ -1,8 +1,6 @@
 defmodule Absinthe.Execution.ArgumentsTest do
   use Absinthe.Case, async: true
 
-  import AssertResult
-
   defmodule Schema do
     use Absinthe.Schema
