+++ test/lib/absinthe/schema_test.exs
@@ -1,7 +1,6 @@
 defmodule Absinthe.SchemaTest do
   use Absinthe.Case, async: true
   use SupportSchemas
-  import AssertResult
 
   alias Absinthe.Schema
   alias Absinthe.Type
