+++ test/support/case.ex
@@ -3,6 +3,7 @@ defmodule Absinthe.Case do
     quote do
       use ExUnit.Case, unquote(opts)
       import Absinthe.Case.Run
+      import Absinthe.Case.Assertions
     end
   end
 end
